+++ buffyboard/regenerate-screenshots.sh
@@ -7,14 +7,8 @@ executable=$1
 outdir=screenshots
 config=buffyboard-screenshots.conf
 
-themes=(
-    breezy-light
-    breezy-dark
-    pmos-light
-    pmos-dark
-    nord-light
-    nord-dark
-)
+root=$(git rev-parse --show-toplevel)
+themes_c=$root/shared/themes.c
 
 resolutions=(
     # Nokia N900
@@ -33,6 +27,11 @@ resolutions=(
     # BQ Aquaris X Pro (landscape)
     1920x1080
 )
+
+if ! which fbcat > /dev/null 2>&1; then
+    echo "Error: Could not find fbcat" 1>&2
+    exit 1
+fi
 
 if [[ ! -f $executable || ! -x $executable ]]; then
     echo "Error: Could not find executable at $executable" 1>&2
@@ -63,7 +62,7 @@ readme="# Buffyboard themes"$'\n'
 
 clear # Blank the screen
 
-for theme in ${themes[@]}; do
+while read -r theme; do
     write_config $theme
 
     readme="$readme"$'\n'"## $theme"$'\n\n'
@@ -74,8 +73,11 @@ for theme in ${themes[@]}; do
 
         sleep 3 # Wait for UI to render
 
-        ../../fbcat/fbcat /dev/fb0 > "$outdir/$theme-$res.ppm"
-        convert -size $fb_res \
+        fbcat /dev/fb0 > "$outdir/$theme-$res.ppm"
+        kill -15 $pid
+
+        convert \
+            -size $fb_res \
             $outdir/$theme-$res.ppm \
             screenshot-background.png \
             -crop $res+0+0 \
@@ -84,12 +86,11 @@ for theme in ${themes[@]}; do
             "$outdir/$theme-$res.png"
 
         rm "$outdir/$theme-$res.ppm"
-        kill -15 $pid
 
         readme="$readme<img src=\"$theme-$res.png\" alt=\"$res\" height=\"300\"/>"$'\n'
 
         sleep 1 # Delay to prevent terminal mode set / reset interference
     done
-done
+done < <(grep "name =" "$themes_c" | grep -o '".*"' | tr -d '"' | sort)
 
 echo -n "$readme" > "$outdir/README.md"