+++ unl0kr/test/test-uses-fb-backend-if-drm-selected-via-config-but-unavailable.sh
@@ -18,7 +18,7 @@ backend=drm
 EOF
 
 info "Running unl0kr"
-run_unl0kr "$log" "$conf"
+run_unl0kr_async "$log" "$conf"
 
 info "Verifying output"
 if ! grep "Using framebuffer backend" "$log"; then